+++ src/shared/hooks/useLoginMutation.ts
@@ -1,6 +1,6 @@
 import { useQueryClient } from '@tanstack/react-query';
 
-import { tokenStorage } from '@/shared/auth/tokenStorage';
+import { setAccessToken } from '@/shared/api';
 import { useMutate } from '@/shared/query/useMutates';
 import { AuthLoginRequest, AuthLoginResponse } from '@/shared/types/response';
 
@@ -9,7 +9,7 @@ const useLoginMutation = () => {
 
   return useMutate<AuthLoginResponse, { method: 'post'; url: string; data: AuthLoginRequest }>({
     onSuccess: (data) => {
-      tokenStorage.set(String(data.accessToken));
+      setAccessToken(String(data.accessToken));
 
       queryClient.invalidateQueries({ queryKey: ['auth', 'me'], exact: false });
     },
